brain_tumor_net-main/requirements.txt
gunicorn==21.2.0
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
protobuf==3.20.*